fix(doctor): preserve migration order before runtime diagnostics
